Prepa A0.9 #1

+++ Views/Battle.py
@@ -49,38 +49,42 @@ class BattleView(lib.discord.ui.View):
         self.add_item(Special_Button())
 
     async def updateBattle(self, timeout=False, poweroff=False):
-        if hasattr(self, "interaction"): message = await self.interaction.original_response()
-        if hasattr(self, "message"): message = self.message
-        if self.Battle.end or timeout:
+        if self.Battle.EndNotPublished:
+            if hasattr(self, "interaction"): message = await self.interaction.original_response()
+            if hasattr(self, "message"): message = self.message
+            if self.Battle.end or timeout:
 
-            if self.Battle.end:
-                await self.Battle.calculateLoot()
+                self.Battle.EndNotPublished = False
 
-            if self.Battle.stats['attacks_received'] > 0 and poweroff==False:
-                content = lib.Toolbox.get_content_looters(self.Battle)
-                embed = lib.Embed.create_embed_end_battle(self.Battle, timeout)
-                if self.Battle.loots != {}:
-                    view = lib.LootRecapView(self.Battle)
-                else:
-                    view = None
+                if self.Battle.end:
+                    await self.Battle.calculateLoot()
 
-                channel = self.Battle.bot.get_channel(self.Battle.bot.rChannels["loots"])
-                if view is None:
-                    await channel.send(embed=embed, view=view)
-                else:
-                    view.message = await channel.send(content=content, embed=embed, view=view)
-            
-            await message.delete()
+                if self.Battle.stats['attacks_received'] > 0 and poweroff==False:
+                    content = lib.Toolbox.get_content_looters(self.Battle)
+                    embed = lib.Embed.create_embed_end_battle(self.Battle, timeout)
+                    if self.Battle.loots != {}:
+                        view = lib.LootRecapView(self.Battle)
+                    else:
+                        view = None
 
-            if not poweroff:
-                #On remove le combat de la liste
-                self.Battle.bot.ActiveList.remove_battle(message.id)
+                    channel = self.Battle.bot.get_channel(self.Battle.bot.rChannels["loots"])
+                    if view is None:
+                        await channel.send(embed=embed, view=view)
+                    else:
+                        view.message = await channel.send(content=content, embed=embed, view=view)
+                        self.Battle.bot.ActiveList.add_recap(view.message.id, view)
+                
+                await message.delete()
 
-            self.stop()
-        else:
-            embed = lib.Embed.create_embed_battle(self.Battle)
-            view = self
-            await message.edit(embed=embed, view=view)
+                if not poweroff:
+                    #On remove le combat de la liste
+                    self.Battle.bot.ActiveList.remove_battle(message.id)
+
+                self.stop()
+            else:
+                embed = lib.Embed.create_embed_battle(self.Battle)
+                view = self
+                await message.edit(embed=embed, view=view)
 
     async def on_timeout(self) -> None:
         await self.updateBattle(True)
